+++ src/lib/pages/interact/component/form/ViewArea.tsx
@@ -46,8 +46,9 @@ export const ViewArea = ({
     typeArgs: getAbiInitialData(fn.generic_type_params.length),
     args: getAbiInitialData(fn.params.length),
   });
-
-  const [abiErrors, setAbiErrors] = useState<[string, string][]>([]);
+  const [abiErrors, setAbiErrors] = useState<[string, string][]>([
+    ["form", "initial"],
+  ]);
   const [res, setRes] = useState<JsonDataType>(undefined);
   const [error, setError] = useState<Option<string>>(undefined);